pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memory-engine"
version = "0.1.0"
description = "Memory Engine MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "rich>=13.7",
]

[project.scripts]
memory = "memory_engine.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
